# app/docs/swagger.yaml
definitions:
...
  categorie.Categorie:
    properties:
      categorie_id:
        example: 1
        type: integer
      type:
        example: Electronics
        type: string
    type: object
  item.Item:
    properties:
      categorie:
        $ref: '#/definitions/categorie.Categorie'
      categorie_id:
        example: 1
        type: integer
      item_id:
        example: 1
        type: integer
      name:
        example: Laptop Pro 15
        type: string
      price:
        example: 999.99
        type: number
      tax:
        example: 0.2
        type: number
      tax_amount:
        example: 199.99
        type: number
    type: object
...
info:
  contact: {}
  title: Tili API
  version: "0.1"
paths:
...
  /categorie:
    get:
      description: Retrieves the complete list of categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/categorie.Categorie'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: List categories
      tags:
      - categorie
    post:
      consumes:
      - application/json
      description: Creates a new categorie in the system. Requires Manager level access.
      parameters:
      - description: Categorie payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/categorie.Categorie'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/categorie.Categorie'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Create a categorie
      tags:
      - categorie
  /categorie/{id}:
    delete:
      description: Deletes a categorie from the system via its ID. Requires Manager
        level access.
      parameters:
      - description: Categorie ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Delete a categorie
      tags:
      - categorie
    get:
      description: Retrieves the details of a categorie using its ID
      parameters:
      - description: Categorie ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categorie.Categorie'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Retrieve a categorie
      tags:
      - categorie
    put:
      consumes:
      - application/json
      description: Modifies the information of an existing categorie via its ID. Requires
        Manager level access.
      parameters:
      - description: Categorie ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Categorie update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/categorie.Categorie'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categorie.Categorie'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Update a categorie
      tags:
      - categorie
  /categorie/type/{type}:
    delete:
      description: Deletes a categorie from the system via its type. Requires Manager
        level access.
      parameters:
      - description: Categorie type
        example: Electronics
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Delete a categorie by type
      tags:
      - categorie
    get:
      description: Retrieves the details of a categorie using its type
      parameters:
      - description: Categorie type
        example: Electronics
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categorie.Categorie'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Retrieve a categorie by type
      tags:
      - categorie
...
  /item/categorie/{id}:
    get:
      description: Retrieves the list of items belonging to a specific category using
        the category ID
      parameters:
      - description: Category ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/item.Item'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ProfileToken: []
      summary: Retrieve items by category
      tags:
      - item
...
securityDefinitions:
  AccountToken:
    description: JWT obtenu après login account (POST /account/login)
    in: header
    name: Authorization
    type: apiKey
  ProfileToken:
    description: JWT obtenu après login profil avec PIN (POST /profile/login/pin)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"